+++ src/AudioMetadataManager.UI/Services/Simulation/Application/Writing/TagLibIntegration/Base/TagLibMetadataWriterBase.cs
@@ -199,7 +199,7 @@ public abstract class TagLibMetadataWriterBase
 
                     preparedFields.Add(
                         _fieldMapper.PrepareChange(
-                            tag,
+                            tagFile,
                             change));
                 }